requirements.txt
streamlit
Pillow
torch
transformers
easyocr
numpy
qwen-vl-utils
byaldi
opencv-python-headless
